backend/requirements.txt
Django==4.2.0
environs==9.5.0
psycopg[binary]==3.1.8 
djangorestframework==3.14.0
